fix(frontend): de-block root layout via suspense boot splash and ios startup images
