+++ frontend/src/app/layout.tsx
@@ -2,18 +2,81 @@ import { SpeedInsights } from "@vercel/speed-insights/next";
 import type { Metadata, Viewport } from "next";
 import { headers } from "next/headers";
 import type { ReactNode } from "react";
-import { AppShell } from "@/components/nav/app-shell";
+import { Suspense } from "react";
+import { AuthShell } from "@/components/auth-shell";
+import { BootSplash } from "@/components/boot-splash";
 import { AppleInstallHint } from "@/components/pwa/apple-install-hint";
 import { SwRegister } from "@/components/pwa/sw-register";
-import { Toaster } from "@/components/ui/sonner";
-import { createSupabaseServerClient } from "@/lib/supabase/server";
 import { Providers } from "./providers";
 import "./globals.css";
 
 export const metadata: Metadata = {
   title: "flamingo-armond",
   description: "Swiping flashcard app.",
-  appleWebApp: { capable: true, statusBarStyle: "default", title: "flamingo" },
+  appleWebApp: {
+    capable: true,
+    statusBarStyle: "default",
+    title: "flamingo",
+    startupImage: [
+      {
+        url: "/splash/splash-1320x2868.png",
+        media: "(device-width:440px) and (device-height:956px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-1206x2622.png",
+        media: "(device-width:402px) and (device-height:874px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-1290x2796.png",
+        media: "(device-width:430px) and (device-height:932px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-1179x2556.png",
+        media: "(device-width:393px) and (device-height:852px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-1170x2532.png",
+        media: "(device-width:390px) and (device-height:844px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-1284x2778.png",
+        media: "(device-width:428px) and (device-height:926px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-1242x2688.png",
+        media: "(device-width:414px) and (device-height:896px) and (-webkit-device-pixel-ratio:3)",
+      },
+      {
+        url: "/splash/splash-828x1792.png",
+        media: "(device-width:414px) and (device-height:896px) and (-webkit-device-pixel-ratio:2)",
+      },
+      {
+        url: "/splash/splash-750x1334.png",
+        media: "(device-width:375px) and (device-height:667px) and (-webkit-device-pixel-ratio:2)",
+      },
+      {
+        url: "/splash/splash-2048x2732.png",
+        media:
+          "(device-width:1024px) and (device-height:1366px) and (-webkit-device-pixel-ratio:2)",
+      },
+      {
+        url: "/splash/splash-1668x2388.png",
+        media: "(device-width:834px) and (device-height:1194px) and (-webkit-device-pixel-ratio:2)",
+      },
+      {
+        url: "/splash/splash-1668x2224.png",
+        media: "(device-width:834px) and (device-height:1112px) and (-webkit-device-pixel-ratio:2)",
+      },
+      {
+        url: "/splash/splash-1488x2266.png",
+        media: "(device-width:744px) and (device-height:1133px) and (-webkit-device-pixel-ratio:2)",
+      },
+      {
+        url: "/splash/splash-1620x2160.png",
+        media: "(device-width:810px) and (device-height:1080px) and (-webkit-device-pixel-ratio:2)",
+      },
+    ],
+  },
 };
 
 export const viewport: Viewport = {
@@ -39,60 +102,6 @@ export default async function RootLayout({ children }: { children: ReactNode }) 
     );
   }
 
-  const supabase = await createSupabaseServerClient();
-  const {
-    data: { user },
-    error,
-  } = await supabase.auth.getUser();
-
-  // AuthSessionMissingError is the normal anonymous-request signal — anything
-  // else from getUser() is a real failure that must degrade the shell to null.
-  const getUserFailed = error != null && error.name !== "AuthSessionMissingError";
-  if (getUserFailed) {
-    console.error("[layout] getUser() failed:", error.name, error.message);
-  }
-
-  // Skip the getClaims call when the user is anonymous: no session means no
-  // JWT to inspect. The /admin layout redirects on actual navigation, so the
-  // header degrades to isAdmin=false for anonymous visitors without noise.
-  let isAdmin = false;
-  if (user && !getUserFailed) {
-    const { data: claimsData, error: claimsError } = await supabase.auth.getClaims();
-    if (claimsError != null) {
-      // getClaims failures are non-fatal: the Supabase session is valid but
-      // the JWT could not be validated locally (e.g. key rotation gap, cold
-      // serverless instance without a cached JWKS). Degrade to isAdmin=false
-      // and warn so the operator can correlate with JWT key rotation events.
-      console.warn("[layout] getClaims() failed", {
-        user_id: user.id,
-        error_name: claimsError.name,
-      });
-    } else if (claimsData == null) {
-      // Third return shape from the SDK: `{ data: null, error: null }` — the
-      // TOCTOU race window where the session vanished between getUser() and
-      // getClaims(). Degrade to isAdmin=false and warn so the operator can
-      // correlate with session-revocation or sign-out events.
-      console.warn("[layout] getClaims() returned null data without error", {
-        user_id: user.id,
-      });
-    } else if (claimsData.claims == null) {
-      // Fourth return shape that is not in the SDK's current TS types but is
-      // forward-compatibility / mock-robustness defence: `{ data: { claims:
-      // null }, error: null }`. The optional chain `.claims?.app_metadata` used
-      // to silently degrade through this case; an explicit narrowing surfaces
-      // it via a warn so an operator can correlate with future SDK type drift.
-      console.warn("[layout] getClaims() returned data without claims", {
-        user_id: user.id,
-      });
-    } else {
-      isAdmin = claimsData.claims.app_metadata?.role === "admin";
-    }
-  }
-
-  // Degrade to null user on a real getUser() error so AppShell renders the
-  // anonymous shell rather than an authenticated shell with potentially stale data.
-  const shellUser = getUserFailed || !user ? null : { email: user.email ?? null };
-
   return (
     <html lang="en">
       {/*
@@ -103,16 +112,9 @@ export default async function RootLayout({ children }: { children: ReactNode }) 
       */}
       <body suppressHydrationWarning>
         <Providers>
-          <AppShell user={shellUser} isAdmin={isAdmin}>
-            {children}
-            {/*
-              Toaster must live inside AppShell (a client-boundary component) because
-              sonner requires a client rendering context. Placing it here ensures the
-              toast container is mounted for every authenticated route while remaining
-              inside the client boundary. RSC + sonner is incompatible at the layout level.
-            */}
-            <Toaster />
-          </AppShell>
+          <Suspense fallback={<BootSplash />}>
+            <AuthShell>{children}</AuthShell>
+          </Suspense>
         </Providers>
         <SpeedInsights />
         <SwRegister />
